Amend FindCommand to support find by tag
FindCommand is amended to extend the current functionality of the find command - allowing users to also find by tags.

Add TagContainsKeywordsPredicate and TagContainsKeywordsPredicateTest to support the extension of the current functionality. Amend the following classes: FindCommand, FindCommandParser to include the use of TagContainsKeywordPredicate. Amend other dependent classes to accommodate the amendments.